Split parents-gate monolith, consolidate shared helpers
