+++ test/lib/server-parents-discovery.test.ts
@@ -20,6 +20,7 @@ import {
   deactivateTestListing,
   describeWithEnv,
   makeParent,
+  makeRoomySharedChild,
   mockRequest,
   publicBody,
   ticketPageStatus,
@@ -289,25 +290,7 @@ describeWithEnv(
         // (B = 1): one parent+child order needs two of A's ten spots, so it fits and
         // the parent keeps its Book link. The pre-fix code took the child's
         // per-listing minimum (1) and marked the parent sold out.
-        const groupA = await createTestGroup({
-          maxAttendees: 10,
-          name: "Shared",
-        });
-        const groupB = await createTestGroup({
-          maxAttendees: 1,
-          name: "Tighter",
-        });
-        const parent = await createTestListing({
-          groupIds: [groupA.id],
-          maxAttendees: 100,
-          name: "Base unit",
-        });
-        const child = await createTestListing({
-          groupIds: [groupA.id, groupB.id],
-          maxAttendees: 100,
-          name: "Add-on",
-        });
-        await listingChildren.setIds(parent.id, [child.id]);
+        const { parent } = await makeRoomySharedChild();
         await assertBookable(parent.slug);
       });
